fix: updated the session values
